+++ visual-js/visual-cypress/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @saucelabs/cypress-visual-plugin
 
+## 0.8.2
+
+### Patch Changes
+
+- Updated dependencies [2d4c4ca]
+  - @saucelabs/visual@0.13.0
+
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies [d6f02a8]
+  - @saucelabs/visual@0.12.0
+
 ## 0.7.0
 
 ### Minor Changes