+++ visual-js/visual-playwright/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @saucelabs/visual-playwright
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [2d4c4ca]
+  - @saucelabs/visual@0.13.0
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [d6f02a8]
+  - @saucelabs/visual@0.12.0
+
 ## 0.2.0
 
 ### Minor Changes