+++ visual-js/visual-storybook/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @saucelabs/visual-storybook
 
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies [2d4c4ca]
+  - @saucelabs/visual@0.13.0
+  - @saucelabs/visual-playwright@0.3.2
+
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies [d6f02a8]
+  - @saucelabs/visual@0.12.0
+  - @saucelabs/visual-playwright@0.3.1
+
 ## 0.9.0
 
 ### Minor Changes